Solve the leak of SockCache entries (#38)
The user-space SockCache must be resilient to the situation of eBPF events
occasionally being missed for tracked sockets.  The NFM process must not grind
to a halt under these conditions.  This change addresses issues within the
SockCache that lead to cache entries being leaked – where they're never evicted
– which causes the NFM process to eventually stop reporting stats.

The issues fixed are: 1) When the SockCache misses a socket close event from
BPF (for example, when the BPF "sk_stats" map is full), or when the SockCache
discards a BPF event (when at `max_sock_entries` capacity), the SockCache fails
to evict entries in a timely manner, and instead waits for the 65 second
staleness timer.  2) The staleness eviction of SockCache entries only considers
sockets for which "sk_stats" events have been received, and does not also
consider sockets for which only a "sk_props" message has been received.

The fixes are: 1) For SockCache to not discard "sk_stats"; 2) To also consider
sockets without stats as eviction candidates; 3) To rapidly evict any sock